move xero error views
